Refactor Timodoro modules
